for preventing import cycle, reorganized defined structs
