+++ app/data/data.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/go-redis/redis/v8"
-	"github.com/itzmeanjan/ette/app/db"
 	"github.com/lib/pq"
 	"gorm.io/gorm"
 )
@@ -253,25 +252,4 @@ func (e *Events) ToJSON() []byte {
 
 	return data
 
-}
-
-// PackedTransaction - All data that is stored in a tx, to be passed from
-// tx data fetcher to whole block data persist handler function
-type PackedTransaction struct {
-	Tx     db.Transactions
-	Events []*db.Events
-}
-
-// PackedTransactions - All tx(s) present in specific block, to be passed
-// using this structure to block data persist handler function
-type PackedTransactions struct {
-	Txs []*PackedTransaction
-}
-
-// PackedBlock - Whole block data to be persisted in a single
-// database transaction to ensure data consistency, if something
-// goes wrong in mid, whole persisting operation will get reverted
-type PackedBlock struct {
-	Block db.Blocks
-	Txs   *PackedTransactions
 }
